<fix>[mevoco]: split LicenseAuthorizedNodeVO.nodeUuid and appId
Related: ZSV-8949

DBImpact
