+++ sdk/src/main/java/org/zstack/sdk/license/api/server/RegisterLicenseServerResult.java
@@ -1,14 +1,23 @@
 package org.zstack.sdk.license.api.server;
 
-
+import org.zstack.sdk.license.header.server.LicenseAuthorizedNodeInventory;
+import org.zstack.sdk.license.header.server.LicenseAuthorizedNodeInventory;
 
 public class RegisterLicenseServerResult {
-    public java.lang.String appId;
-    public void setAppId(java.lang.String appId) {
-        this.appId = appId;
+    public LicenseAuthorizedNodeInventory licenseClient;
+    public void setLicenseClient(LicenseAuthorizedNodeInventory licenseClient) {
+        this.licenseClient = licenseClient;
     }
-    public java.lang.String getAppId() {
-        return this.appId;
+    public LicenseAuthorizedNodeInventory getLicenseClient() {
+        return this.licenseClient;
+    }
+
+    public LicenseAuthorizedNodeInventory licenseServer;
+    public void setLicenseServer(LicenseAuthorizedNodeInventory licenseServer) {
+        this.licenseServer = licenseServer;
+    }
+    public LicenseAuthorizedNodeInventory getLicenseServer() {
+        return this.licenseServer;
     }
 
 }
